paperenvbench/tasks/dinov2_feature_minimal/requirements_lock.txt
-e git+https://github.com/facebookresearch/dinov2.git@7b187bd4df8efce2cbcbbb67bd01532c19bf4c9c#egg=dinov2
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.3
antlr4-python3-runtime==4.9.3
filelock==3.29.0
fsspec==2026.4.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
packaging==26.2
pillow==12.1.0
setuptools==82.0.1
sympy==1.14.0
torch==2.8.0+cpu
torchvision==0.23.0+cpu
typing_extensions==4.15.0
wheel==0.47.0
